+++ R/BarDiscrete.r
@@ -35,7 +35,7 @@ BarDiscrete <- function(data, na.rm = TRUE, maxcat = 50) {
   ign_ind <- which(n_cat > maxcat)
   if (length(ign_ind) > 0) {
     set(discrete, j = ign_ind, value = NULL)
-    cat(length(ign_ind), "columns ignored with more than", maxcat, "categories.\n", paste0(names(ign_ind), ": ", n_cat[ign_ind], " categories\n"))
+    message(length(ign_ind), " columns ignored with more than ", maxcat, " categories.\n", paste0(names(ign_ind), ": ", n_cat[ign_ind], " categories\n"))
   }
   ## Get dimension
   n <- nrow(discrete)