+++ NEWS.md
@@ -5,10 +5,12 @@
 ### DataExplorer 0.3.0.9000
 #### Bug Fixes
 * [#32](https://github.com/boxuancui/DataExplorer/issues/32): Fixed pandoc requirement error in unit test on cran.
-* [#34](https://github.com/boxuancui/DataExplorer/issues/34): Fixed error message when `quiet` is not supplied. In addition, report directory are printed through `message()` instead of `cat()`. 
+* [#34](https://github.com/boxuancui/DataExplorer/issues/34): Fixed error message when `quiet` is not supplied. In addition, report directory are printed through `message()` instead of `cat()`.
+* [#35](https://github.com/boxuancui/DataExplorer/issues/35): Added **rprojroot** to Suggests.
 
 #### Enhancements
 * [#36](https://github.com/boxuancui/DataExplorer/issues/36): Fixed warnings from data.table in `DropVar`.
+* [#37](https://github.com/boxuancui/DataExplorer/issues/37): Changed all `cat()` to `message()`.
 * Added more examples in README file.
 
 ---