+++ NEWS.md
@@ -5,11 +5,12 @@
 ### DataExplorer 0.3.0.9000
 #### New Features
 * [#33](https://github.com/boxuancui/DataExplorer/issues/33): Added `PlotStr` to visualize data structure.
+* [#40](https://github.com/boxuancui/DataExplorer/issues/40): Added network graph to `GenerateReport`.
 
 #### Bug Fixes
 * [#32](https://github.com/boxuancui/DataExplorer/issues/32): Fixed pandoc requirement error in unit test on cran.
 * [#34](https://github.com/boxuancui/DataExplorer/issues/34): Fixed error message when `quiet` is not supplied. In addition, report directory are printed through `message()` instead of `cat()`.
-* [#35](https://github.com/boxuancui/DataExplorer/issues/35): Added **rprojroot** to Suggests.
+* [#35](https://github.com/boxuancui/DataExplorer/issues/35): Fixed **rprojroot** not found error.
 
 #### Enhancements
 * [#36](https://github.com/boxuancui/DataExplorer/issues/36): Fixed warnings from data.table in `DropVar`.